Rename helpers to formatters.
